Move 'renderer.RenderData, render_data_utils' into 'renderer.render_data'
