+++ src/main/scala/engine/render/renderer/DefaultRenderer.scala
@@ -12,6 +12,7 @@ import org.lwjgl.opengl.GL20._
 import org.lwjgl.opengl.GL30._
 import engine.math.Vector3
 import engine.math.Vector4
+import render_data.RenderData
 
 final case class DefaultRenderer(
     override val window: Window
